refactor: 메인페이지 추상화
- MainPage 컴포넌트 내부를 추상화하기 위해 MainFeatureMenuContainer라는 컴포넌트를 만들어주었다.
- page 컴포넌트 컨벤션에 따라서, 이렇게 진행
